Observer Pattern implemented to handle errors
The most significant changes involve the addition of an `OnError` event to the `IIngredientsRepository` interface and its implementation in the `IngredientsRepository` class. This event is triggered when an error occurs, passing the error message as a parameter. The `IngredientsForm` class now subscribes to this event, displaying the error message in a message box and setting the `_errorOccured` flag to `true` when the event is triggered.

Additionally, error handling in the `AddIngredient` method of the `IngredientsRepository` class has been refactored to use a new `ErrorOccured` method. This method triggers the `OnError` event and logs the error.

The `AddToKeteBtn_Click` method in the `IngredientsForm` class has been updated to clear all fields only if no error occurred. If an error occurred, only the input fields are cleared.

Finally, the check for duplicate ingredient names in the `IsValid` method of the `IngredientsForm` class has been removed, as this is now handled by the `OnError` event.

List of changes:

1. Addition of `OnError` event to `IIngredientsRepository` interface and `IngredientsRepository` class. This event is triggered when an error occurs, passing the error message as a parameter.
2. Addition of `ErrorOccured` method to `IngredientsRepository` class. This method triggers the `OnError` event and logs the error.
3. Refactoring of error handling in `AddIngredient` method of `IngredientsRepository` class to use `ErrorOccured` method.
4. `IngredientsForm` class now subscribes to `OnError` event of `IIngredientsRepository` instance. When event is triggered, `OnErrorOccured` method is called, displaying error message in a message box and setting `_errorOccured` flag to `true`.
5. Update to `AddToKeteBtn_Click` method in `IngredientsForm` class to clear all fields only if no error occurred. If error occurred, only input fields are cleared.
6. Removal of check for duplicate ingredient names in `IsValid` method of `IngredientsForm` class, as this is now handled by `OnError` event.
7. Comment added to `Logger` class indicating that other types of logs should be implemented in addition to error logs.